+++ src/Worker.js
@@ -20,6 +20,13 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import { combinations } from './MathUtils.js';
 
 const PHYSICAL = 1;
+const FIRE = 1;
+const WATER = 2;
+const WIND = 3;
+const SWORD = 5;
+const HAMMER = 6;
+const BOW = 7;
+const POLE = 8;
 
 const evaluateDamagePerSP = (deck, playerStats, options, keys) => {
   let totalCost = 0;
@@ -29,6 +36,16 @@ const evaluateDamagePerSP = (deck, playerStats, options, keys) => {
   let totalPhysicalWeaponMult = 0;
   let totalMagicalWeaponMult = 0;
   let supportSkillMult = 0;
+
+  let elementCount = [];
+  elementCount[FIRE] = 0;
+  elementCount[WATER] = 0;
+  elementCount[WIND] = 0;
+  let typeCount = [];
+  typeCount[SWORD] = 0;
+  typeCount[HAMMER] = 0;
+  typeCount[BOW] = 0;
+  typeCount[POLE] = 0;
 
   keys.forEach((k) => {
     const weapon = deck[k];
@@ -43,16 +60,27 @@ const evaluateDamagePerSP = (deck, playerStats, options, keys) => {
       totalMagicalWeaponMult += weapon.skill_mult;
     }
     supportSkillMult += weapon.supp_skill_mult;
+
+    elementCount[weapon.element]++;
+    typeCount[weapon.type]++;
   });
 
   if (totalCost > playerStats.maxCost) {
     return 0;
   }
 
-  const enemyPDef = options.targetPDef;
-  const enemyMDef = options.targetMDef;
-  const effectivePAtk = Math.max(0, totalPhysicalAttack - (2 * enemyPDef / 3));
-  const effectiveMAtk = Math.max(0, totalMagicalAttack - (2 * enemyMDef / 3));
+  if (elementCount[FIRE] < options.minFire ||
+      elementCount[WATER] < options.minWater ||
+      elementCount[WIND] < options.minWind ||
+      typeCount[SWORD] < options.minSwords ||
+      typeCount[HAMMER] < options.minHammers ||
+      typeCount[BOW] < options.minBows ||
+      typeCount[POLE] < options.minPoles) {
+    return 0;
+  }
+
+  const effectivePAtk = Math.max(0, totalPhysicalAttack * options.pAtkCorrection - options.effectiveEnemyPDef);
+  const effectiveMAtk = Math.max(0, totalMagicalAttack * options.mAtkCorrection - options.effectiveEnemyMDef);
 
   const damage = (effectivePAtk * totalPhysicalWeaponMult + effectiveMAtk * totalMagicalWeaponMult) * (1 + supportSkillMult);
   return damage / totalSPCost;
@@ -66,6 +94,16 @@ const evaluateDamage = (deck, playerStats, options, keys) => {
   let totalMagicalWeaponMult = 0;
   let supportSkillMult = 1;
 
+  let elementCount = [];
+  elementCount[FIRE] = 0;
+  elementCount[WATER] = 0;
+  elementCount[WIND] = 0;
+  let typeCount = [];
+  typeCount[SWORD] = 0;
+  typeCount[HAMMER] = 0;
+  typeCount[BOW] = 0;
+  typeCount[POLE] = 0;
+
   keys.forEach((k) => {
     const weapon = deck[k];
 
@@ -78,16 +116,27 @@ const evaluateDamage = (deck, playerStats, options, keys) => {
       totalMagicalWeaponMult += weapon.skill_mult;
     }
     supportSkillMult += weapon.supp_skill_mult;
+
+    elementCount[weapon.element]++;
+    typeCount[weapon.type]++;
   });
 
   if (totalCost > playerStats.maxCost) {
     return 0;
   }
 
-  const enemyPDef = options.targetPDef;
-  const enemyMDef = options.targetMDef;
-  const effectivePAtk = totalPhysicalAttack - (2 * enemyPDef / 3);
-  const effectiveMAtk = totalMagicalAttack - (2 * enemyMDef / 3);
+  if (elementCount[FIRE] < options.minFire ||
+      elementCount[WATER] < options.minWater ||
+      elementCount[WIND] < options.minWind ||
+      typeCount[SWORD] < options.minSwords ||
+      typeCount[HAMMER] < options.minHammers ||
+      typeCount[BOW] < options.minBows ||
+      typeCount[POLE] < options.minPoles) {
+    return 0;
+  }
+
+  const effectivePAtk = Math.max(0, totalPhysicalAttack * options.pAtkCorrection - options.effectiveEnemyPDef);
+  const effectiveMAtk = Math.max(0, totalMagicalAttack * options.mAtkCorrection - options.effectiveEnemyMDef);
 
   const damage = (effectivePAtk * totalPhysicalWeaponMult + effectiveMAtk * totalMagicalWeaponMult) * (1 + supportSkillMult);
   return damage;